# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/me/clear_data": {
            "delete": {
                "description": "Очистить все пользовательские данные.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "Очистить все пользовательские данные.",
                "responses": {
                    "200": {
                        "description": "success clear user data"
                    },
                    "500": {
                        "description": "internal server error",
                        "schema": {
                            "$ref": "#/definitions/echo.HTTPError"
                        }
                    }
                }
            }
        },
...
        "/me/projects": {
            "get": {
                "description": "Получить список проектов пользователя.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "projects"
                ],
                "summary": "Получить список проектов.",
                "responses": {
                    "200": {
                        "description": "success get projects",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/internal_project_delivery.ProjectOut"
                            }
                        }
                    },
                    "400": {
                        "description": "bad request",
                        "schema": {
                            "$ref": "#/definitions/echo.HTTPError"
                        }
                    },
                    "500": {
                        "description": "internal server error",
                        "schema": {
                            "$ref": "#/definitions/echo.HTTPError"
                        }
                    }
                }
            }
        },
        "/me/projects/stat": {
            "get": {
                "description": "Получить статистику по проектам.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "projects"
                ],
                "summary": "Получить статистику по проектам.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "RFC3339 format",
                        "name": "time_start",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "RFC3339 format",
                        "name": "time_end",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "schema": {
                            "$ref": "#/definitions/internal_project_delivery.ProjectsStatOut"
                        }
                    },
                    "500": {
                        "description": "internal server error",
                        "schema": {
                            "$ref": "#/definitions/echo.HTTPError"
                        }
                    }
                }
            }
        },
        "/me/projects/{id}/stat": {
            "get": {
                "description": "Получить статистику по конкретному проекту.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "projects"
                ],
                "summary": "Получить статистику по конкретному проекту.",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "project ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "RFC3339 format",
                        "name": "time_start",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "RFC3339 format",
                        "name": "time_end",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "schema": {
                            "$ref": "#/definitions/internal_project_delivery.ProjectEntriesStatOut"
                        }
                    },
                    "400": {
                        "description": "bad request",
                        "schema": {
                            "$ref": "#/definitions/echo.HTTPError"
                        }
                    },
                    "500": {
                        "description": "internal server error",
                        "schema": {
                            "$ref": "#/definitions/echo.HTTPError"
                        }
                    }
                }
            }
        },
...
        "/projects/create": {
            "post": {
                "description": "Создать проект.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "projects"
                ],
                "summary": "Создать проект.",
                "parameters": [
                    {
                        "description": "project info",
                        "name": "project",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_project_delivery.CreateProjectIn"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success create project",
                        "schema": {
                            "$ref": "#/definitions/internal_project_delivery.CreateProjectOut"
                        }
                    },
                    "400": {
                        "description": "bad request",
                        "schema": {
                            "$ref": "#/definitions/echo.HTTPError"
                        }
                    },
                    "422": {
                        "description": "unprocessable entity",
                        "schema": {
                            "$ref": "#/definitions/echo.HTTPError"
                        }
                    },
                    "500": {
                        "description": "internal server error",
                        "schema": {
                            "$ref": "#/definitions/echo.HTTPError"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "echo.HTTPError": {
            "type": "object",
            "properties": {
                "message": {}
            }
        },
...
        "internal_project_delivery.CreateProjectIn": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "name": {
                    "description": "Название проекта.",
                    "type": "string",
                    "example": "Работа"
                }
            }
        },
        "internal_project_delivery.CreateProjectOut": {
            "type": "object",
            "required": [
                "id"
            ],
            "properties": {
                "id": {
                    "description": "Идентификатор проекта.",
                    "type": "integer",
                    "example": 1
                }
            }
        },
        "internal_project_delivery.ProjectEntriesStat": {
            "type": "object",
            "properties": {
                "duration_in_sec": {
                    "description": "Суммарное время (в сек.) потраченное на запись.",
                    "type": "number",
                    "example": 360
                },
                "name": {
                    "description": "Название записи.",
                    "type": "string",
                    "example": "task1"
                },
                "percent_duration": {
                    "description": "Доля (в процентах) длительности записи от длительности проекта.",
                    "type": "number"
                }
            }
        },
        "internal_project_delivery.ProjectEntriesStatOut": {
            "type": "object",
            "properties": {
                "entries": {
                    "description": "Записи времени.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/internal_project_delivery.ProjectEntriesStat"
                    }
                },
                "total_duration_in_sec": {
                    "description": "Суммарное время (в сек.) потраченное на проект.",
                    "type": "number",
                    "example": 60
                }
            }
        },
        "internal_project_delivery.ProjectOut": {
            "type": "object",
            "properties": {
                "id": {
                    "description": "Идентификатор проекта.                         // Идентификатор проекта.",
                    "type": "integer",
                    "example": 1
                },
                "name": {
                    "description": "Название проекта.",
                    "type": "string",
                    "example": "Работа"
                }
            }
        },
        "internal_project_delivery.ProjectStat": {
            "type": "object",
            "properties": {
                "duration_in_sec": {
                    "description": "Суммарное время (в сек.) потраченное на проект.",
                    "type": "number",
                    "example": 360
                },
                "id": {
                    "description": "Идентификатор проекта.",
                    "type": "integer",
                    "example": 1
                },
                "name": {
                    "description": "Название проекта.",
                    "type": "string",
                    "example": "Работа"
                },
                "percent_duration": {
                    "description": "Доля (в процентах) длительности проекта от суммарной длительности.",
                    "type": "number",
                    "example": 10
                }
            }
        },
        "internal_project_delivery.ProjectsStatOut": {
            "type": "object",
            "properties": {
                "projects": {
                    "description": "Список проектов.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/internal_project_delivery.ProjectStat"
                    }
                },
                "total_duration_in_sec": {
                    "description": "Суммарное время (в сек.) потраченное на все проекты.",
                    "type": "number",
                    "example": 3600
                }
            }
        }
    }
}